requirements.txt
instagrapi
requests
Flask
flask-cors
python-dotenv
gunicorn
psycopg2-binary
